Inline _training.py back into training.py
Drop the thin re-export wrapper — there is only one module now, so the
private/public split adds noise without a payoff. The branch on
matgl.config.BACKEND lives directly in training.py.

Both backends' tests still pass (PyG 19/19, DGL 18/18); ruff, ruff format,
and mypy are clean.
